# `defendpoint` 2.1: 3x performance boost