--- a/.clj-kondo/test/hooks/metabase/api/macros_test.clj
+++ b/.clj-kondo/test/hooks/metabase/api/macros_test.clj
@@ -4,6 +4,11 @@
    [clojure.test :refer :all]
    [hooks.metabase.api.macros]))
 
+(defn- is-valid-node [x]
+  (is (api/node? x))
+  (doseq [child (:children x)]
+    (is-valid-node child)))
+
 (deftest ^:parallel defendpoint-test
   (let [form '(api.macros/defendpoint :post "/" :- [:map [:collection_id :int]]
                 "Create a new [[Timeline]]."
@@ -15,21 +20,23 @@
                 (body icon)
                 (body collection-id))
         node (-> form pr-str api/parse-string)]
-    (is (= '(do
-              api.macros/defendpoint
-              :post
-              "/"
-              (do
-                [:map [:collection_id :int]]
-                [:map
-                 [:name    ms/NonBlankString]
-                 [:default {:optional true} [:maybe :boolean]]])
-              (clojure.core/let [_route-params nil
-                                 _query-params nil
-                                 {:keys [icon], collection-id :collection_id, :as body} nil]
-                (body icon)
-                (body collection-id)))
-           (-> {:node node}
-               hooks.metabase.api.macros/defendpoint
-               :node
-               api/sexpr)))))
+    (is-valid-node node)
+    (let [node' (-> {:node node}
+                    hooks.metabase.api.macros/defendpoint
+                    :node)]
+      (is-valid-node node')
+      (is (= '(do
+                api.macros/defendpoint
+                :post
+                "/"
+                (do
+                  [:map [:collection_id :int]]
+                  [:map
+                   [:name    ms/NonBlankString]
+                   [:default {:optional true} [:maybe :boolean]]])
+                (clojure.core/let [_route-params {}
+                                   _query-params {}
+                                   {:keys [icon], collection-id :collection_id, :as body} {}]
+                  (body icon)
+                  (body collection-id)))
+             (api/sexpr node'))))))