--- a/src/metabase/api/common.clj
+++ b/src/metabase/api/common.clj
@@ -439,14 +439,10 @@
     ;; for dev, fetch the handler from the metadata on every request so we get nice live reloading if the endpoints in a
     ;; namespace change.
     (if config/is-dev?
-      (fn
-        ([request]
-         (when-let [handler (handler-fn)]
-           (handler request)))
-        ([request respond raise]
-         (if-let [handler (handler-fn)]
-           (handler request respond raise)
-           (respond nil))))
+      (fn [request respond raise]
+        (if-let [handler (handler-fn)]
+          (handler request respond raise)
+          (respond nil)))
       ;; For prod, fetching the handler on each request gives us nothing since it shouldn't change; fetch it once and if
       ;; it's not defined just use the [[pass-thru-handler]] above instead.
       (or (handler-fn) pass-thru-handler))))