- Change the `&request` anaphor introduced in #52570 to an optional 4th `request` parameter -- this lets us do Malli schema validation and decoding on the request as well
- Support `{:mulitpart true}` requests (from #52642)
- Support optional 5th and 6th `respond` and `raise` args for writing async handlers
- Update the endpoints in `geojson.clj` to use defendpoint 2
- Vastly more efficient compiled combined handler for a namespace finds correct matching handler and then calls it as opposed to trying every single handler one by one and then calling a fallback function -- **this also results in stacktraces that are dramatically shorter when API calls fail**. Combined handler also avoids a lot of unnecessary Compojure overhead that is unneeded or unused by our endpoints
- Eliminate need for separate `base-handler` and `route-handler` -- simpler macroexpansion 

## `defendpoint` 2 confirmed to be 3x faster than `defendpoint` 1

<details>

<summary>Methodology</summary>

I created 52 API endpoints like these with both defendpoint 1 and 2 in two separate namespaces:

```clj
(api/defendpoint GET  "/a/:id" [id] {id pos-int?} {:a id})
(api/defendpoint POST "/a/:id" [id] {id pos-int?} {:a id})
(api/defendpoint GET  "/b/:id" [id] {id pos-int?} {:b id})
(api/defendpoint POST "/b/:id" [id] {id pos-int?} {:b id})
(api/defendpoint GET  "/c/:id" [id] {id pos-int?} {:c id})
(api/defendpoint POST "/c/:id" [id] {id pos-int?} {:c id})
(api/defendpoint GET  "/d/:id" [id] {id pos-int?} {:d id})
(api/defendpoint POST "/d/:id" [id] {id pos-int?} {:d id})
(api/defendpoint GET  "/e/:id" [id] {id pos-int?} {:e id})
(api/defendpoint POST "/e/:id" [id] {id pos-int?} {:e id})
(api/defendpoint GET  "/f/:id" [id] {id pos-int?} {:f id})
(api/defendpoint POST "/f/:id" [id] {id pos-int?} {:f id})
(api/defendpoint GET  "/g/:id" [id] {id pos-int?} {:g id})
(api/defendpoint POST "/g/:id" [id] {id pos-int?} {:g id})
(api/defendpoint GET  "/h/:id" [id] {id pos-int?} {:h id})
(api/defendpoint POST "/h/:id" [id] {id pos-int?} {:h id})
(api/defendpoint GET  "/i/:id" [id] {id pos-int?} {:i id})
(api/defendpoint POST "/i/:id" [id] {id pos-int?} {:i id})
(api/defendpoint GET  "/j/:id" [id] {id pos-int?} {:j id})
(api/defendpoint POST "/j/:id" [id] {id pos-int?} {:j id})
(api/defendpoint GET  "/k/:id" [id] {id pos-int?} {:k id})
(api/defendpoint POST "/k/:id" [id] {id pos-int?} {:k id})
(api/defendpoint GET  "/l/:id" [id] {id pos-int?} {:l id})
(api/defendpoint POST "/l/:id" [id] {id pos-int?} {:l id})
(api/defendpoint GET  "/m/:id" [id] {id pos-int?} {:m id})
(api/defendpoint POST "/m/:id" [id] {id pos-int?} {:m id})
(api/defendpoint GET  "/n/:id" [id] {id pos-int?} {:n id})
(api/defendpoint POST "/n/:id" [id] {id pos-int?} {:n id})
(api/defendpoint GET  "/o/:id" [id] {id pos-int?} {:o id})
(api/defendpoint POST "/o/:id" [id] {id pos-int?} {:o id})
(api/defendpoint GET  "/p/:id" [id] {id pos-int?} {:p id})
(api/defendpoint POST "/p/:id" [id] {id pos-int?} {:p id})
(api/defendpoint GET  "/q/:id" [id] {id pos-int?} {:q id})
(api/defendpoint POST "/q/:id" [id] {id pos-int?} {:q id})
(api/defendpoint GET  "/r/:id" [id] {id pos-int?} {:r id})
(api/defendpoint POST "/r/:id" [id] {id pos-int?} {:r id})
(api/defendpoint GET  "/s/:id" [id] {id pos-int?} {:s id})
(api/defendpoint POST "/s/:id" [id] {id pos-int?} {:s id})
(api/defendpoint GET  "/t/:id" [id] {id pos-int?} {:t id})
(api/defendpoint POST "/t/:id" [id] {id pos-int?} {:t id})
(api/defendpoint GET  "/u/:id" [id] {id pos-int?} {:u id})
(api/defendpoint POST "/u/:id" [id] {id pos-int?} {:u id})
(api/defendpoint GET  "/v/:id" [id] {id pos-int?} {:v id})
(api/defendpoint POST "/v/:id" [id] {id pos-int?} {:v id})
(api/defendpoint GET  "/w/:id" [id] {id pos-int?} {:w id})
(api/defendpoint POST "/w/:id" [id] {id pos-int?} {:w id})
(api/defendpoint GET  "/x/:id" [id] {id pos-int?} {:x id})
(api/defendpoint POST "/x/:id" [id] {id pos-int?} {:x id})
(api/defendpoint GET  "/y/:id" [id] {id pos-int?} {:y id})
(api/defendpoint POST "/y/:id" [id] {id pos-int?} {:y id})
(api/defendpoint GET  "/z/:id" [id] {id pos-int?} {:z id})
```

then I called the namespace handlers with a mock request like this:

```clj
(def mock-request
  {:accept         "json"
   :headers        {"content-type" "application/json"}
   :request-method :get
   :uri            "/api/b/100"
   :compojure/path "/b/100"})

(defn raise [e] (throw e))

(defn x []
  (println "legacy defendpoint")
  (criterium.core/quick-bench (metabase.api.x/routes mock-request :body raise))

  (println "defendpoint 2")
  (criterium.core/quick-bench (metabase.api.y/routes mock-request :body raise)))
```

I tried this with both `/a/` (which should be the first handler tried) and with `/z` (which should be the last handler tried)

</details>

```
GET /api/a/100 

legacy defendpoint
Evaluation count : 45108 in 6 samples of 7518 calls.
             Execution time mean : 14.430447 µs
    Execution time std-deviation : 1.938135 µs
   Execution time lower quantile : 13.233018 µs ( 2.5%)
   Execution time upper quantile : 17.655664 µs (97.5%)
                   Overhead used : 1.516265 ns

defendpoint 2
Evaluation count : 238866 in 6 samples of 39811 calls.
             Execution time mean : 2.592549 µs
    Execution time std-deviation : 72.868994 ns
   Execution time lower quantile : 2.548722 µs ( 2.5%)
   Execution time upper quantile : 2.714488 µs (97.5%)
                   Overhead used : 1.516265 ns

GET /api/z/100 

legacy defendpoint
Evaluation count : 37098 in 6 samples of 6183 calls.
             Execution time mean : 18.804212 µs
    Execution time std-deviation : 3.387020 µs
   Execution time lower quantile : 16.269834 µs ( 2.5%)
   Execution time upper quantile : 23.285804 µs (97.5%)
                   Overhead used : 1.516265 ns

defendpoint 2
Evaluation count : 128124 in 6 samples of 21354 calls.
             Execution time mean : 6.331762 µs
    Execution time std-deviation : 1.938481 µs
   Execution time lower quantile : 4.639180 µs ( 2.5%)
   Execution time upper quantile : 9.507562 µs (97.5%)
                   Overhead used : 1.516265 ns
```